requirements.txt

reflex==0.8.15a1
pymongo
mysql-connector-python
bson